+++ src/components/common/FocusTrapper.js
@@ -1,0 +1,20 @@
+import _ from 'lodash/fp'
+import FocusLock from 'react-focus-lock'
+import { h } from 'react-hyperscript-helpers'
+
+
+export const FocusTrapper = ({ children, onBreakout, ...props }) => {
+  return h(FocusLock, {
+    returnFocus: true,
+    lockProps: _.merge({
+      tabIndex: 0,
+      style: { outline: 'none' },
+      onKeyDown: e => {
+        if (e.key === 'Escape') {
+          onBreakout()
+          e.stopPropagation()
+        }
+      }
+    }, props)
+  }, [children])
+}
